+++ src/main/java/model/ChatRoom.java
@@ -3,13 +3,13 @@ package model;
 import java.sql.Timestamp;
 
 public class ChatRoom {
-    private long id;
-    private long productId;
-    private long buyerId;
+    private int id;
+    private int productId;
+    private int buyerId;
     private Timestamp createdAt;
 
     // 생성자1: DB 조회 시 (모든 컬럼 포함)
-    public ChatRoom(long id, long productId, long buyerId, Timestamp createdAt) {
+    public ChatRoom(int id, int productId, int buyerId, Timestamp createdAt) {
         this.id = id;
         this.productId = productId;
         this.buyerId = buyerId;
@@ -17,7 +17,7 @@ public class ChatRoom {
     }
 
     // 생성자2: 새로 생성할 때 (createdAt 자동 생성)
-    public ChatRoom(long id, long productId, long buyerId) {
+    public ChatRoom(int id, int productId, int buyerId) {
         this.id = id;
         this.productId = productId;
         this.buyerId = buyerId;
@@ -25,13 +25,13 @@ public class ChatRoom {
     }
 
     // Getter / Setter
-    public long getId() { return id; }
-    public long getProductId() { return productId; }
-    public long getBuyerId() { return buyerId; }
+    public int getId() { return id; }
+    public int getProductId() { return productId; }
+    public int getBuyerId() { return buyerId; }
     public Timestamp getCreatedAt() { return createdAt; }
 
-    public void setId(long id) { this.id = id; }
-    public void setProductId(long productId) { this.productId = productId; }
-    public void setBuyerId(long buyerId) { this.buyerId = buyerId; }
+    public void setId(int id) { this.id = id; }
+    public void setProductId(int productId) { this.productId = productId; }
+    public void setBuyerId(int buyerId) { this.buyerId = buyerId; }
     public void setCreatedAt(Timestamp createdAt) { this.createdAt = createdAt; }
 }
