+++ src/main/java/model/ChatRoomDisplayDTO.java
@@ -3,15 +3,15 @@ package model;
 import java.sql.Timestamp;
 
 public class ChatRoomDisplayDTO {
-    private long id;
-    private long productId;
-    private long buyerId;
+    private int id;
+    private int productId;
+    private int buyerId;
     private Timestamp createdAt;
     private String productTitle;
     private String otherUserNickname;
-    private long otherUserId;
+    private int otherUserId;
 
-    public ChatRoomDisplayDTO(long id, long productId, long buyerId, Timestamp createdAt, String productTitle, String otherUserNickname, long otherUserId) {
+    public ChatRoomDisplayDTO(int id, int productId, int buyerId, Timestamp createdAt, String productTitle, String otherUserNickname, int otherUserId) {
         this.id = id;
         this.productId = productId;
         this.buyerId = buyerId;
@@ -22,20 +22,20 @@ public class ChatRoomDisplayDTO {
     }
 
     // Getters
-    public long getId() { return id; }
-    public long getProductId() { return productId; }
-    public long getBuyerId() { return buyerId; }
+    public int getId() { return id; }
+    public int getProductId() { return productId; }
+    public int getBuyerId() { return buyerId; }
     public Timestamp getCreatedAt() { return createdAt; }
     public String getProductTitle() { return productTitle; }
     public String getOtherUserNickname() { return otherUserNickname; }
-    public long getOtherUserId() { return otherUserId; }
+    public int getOtherUserId() { return otherUserId; }
 
     // Setters (if needed, but for DTOs, often only getters are used)
-    public void setId(long id) { this.id = id; }
-    public void setProductId(long productId) { this.productId = productId; }
-    public void setBuyerId(long buyerId) { this.buyerId = buyerId; }
+    public void setId(int id) { this.id = id; }
+    public void setProductId(int productId) { this.productId = productId; }
+    public void setBuyerId(int buyerId) { this.buyerId = buyerId; }
     public void setCreatedAt(Timestamp createdAt) { this.createdAt = createdAt; }
     public void setProductTitle(String productTitle) { this.productTitle = productTitle; }
     public void setOtherUserNickname(String otherUserNickname) { this.otherUserNickname = otherUserNickname; }
-    public void setOtherUserId(long otherUserId) { this.otherUserId = otherUserId; }
+    public void setOtherUserId(int otherUserId) { this.otherUserId = otherUserId; }
 }
