+++ src/main/java/model/Message.java
@@ -3,9 +3,9 @@ package model;
 import java.sql.Timestamp;
 
 public class Message {
-    private long id;
-    private long chatRoomId;
-    private long senderId;
+    private int id;
+    private int chatRoomId;
+    private int senderId;
     private String message;
     private Timestamp createdAt;
 
@@ -13,7 +13,7 @@ public class Message {
     public Message() {}
 
     // 생성자 (DB 조회용)
-    public Message(long id, long chatRoomId, long senderId, String message, Timestamp createdAt) {
+    public Message(int id, int chatRoomId, int senderId, String message, Timestamp createdAt) {
         this.id = id;
         this.chatRoomId = chatRoomId;
         this.senderId = senderId;
@@ -22,7 +22,7 @@ public class Message {
     }
 
     // 생성자 (새 메시지 저장용)
-    public Message(long chatRoomId, long senderId, String message) {
+    public Message(int chatRoomId, int senderId, String message) {
         this.chatRoomId = chatRoomId;
         this.senderId = senderId;
         this.message = message;
@@ -30,14 +30,14 @@ public class Message {
     }
 
     // Getter / Setter
-    public long getId() { return id; }
-    public void setId(long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
-    public long getChatRoomId() { return chatRoomId; }
-    public void setChatRoomId(long chatRoomId) { this.chatRoomId = chatRoomId; }
+    public int getChatRoomId() { return chatRoomId; }
+    public void setChatRoomId(int chatRoomId) { this.chatRoomId = chatRoomId; }
 
-    public long getSenderId() { return senderId; }
-    public void setSenderId(long senderId) { this.senderId = senderId; }
+    public int getSenderId() { return senderId; }
+    public void setSenderId(int senderId) { this.senderId = senderId; }
 
     public String getMessage() { return message; }
     public void setMessage(String message) { this.message = message; }
